requirements.txt
fastapi~=0.85.1
uvicorn~=0.19.0
pydantic~=1.10.2
sympy~=1.11.1
Pillow~=9.3.0
starlette~=0.20.4
passlib~=1.7.4
bcrypt~=4.0.1
python-jose~=3.3.0
python-multipart~=0.0.5